backend/vnc_bridge/app_real.py
# ABOUTME: Real VNC bridge implementation with proper routing
# ABOUTME: Routes VNC connections to correct desktop containers based on session ID

import os
import asyncio
import base64
import json
import boto3
import websockets
from fastapi import FastAPI, WebSocket, WebSocketDisconnect, HTTPException, Request
from fastapi.responses import HTMLResponse, Response
from fastapi.middleware.cors import CORSMiddleware
import logging
from typing import Optional, Dict
import struct

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# AWS clients
dynamodb = boto3.resource('dynamodb', region_name='us-west-2')
sessions_table = dynamodb.Table(os.environ.get('SESSIONS_TABLE', 'computer-use-sessions-dev'))

app = FastAPI(title="VNC Bridge Service")

# Enable CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Cache for session lookups
session_cache: Dict[str, dict] = {}

async def get_desktop_ip(session_id: str) -> Optional[str]:
    """Get desktop container IP for a session"""
    
    # Check cache first
    if session_id in session_cache:
        return session_cache[session_id].get('task_ip')
    
    try:
        # Look up session in DynamoDB
        response = sessions_table.get_item(Key={'session_id': session_id})
        session = response.get('Item')
        
        if session and session.get('task_ip'):
            session_cache[session_id] = session
            return session['task_ip']
            
    except Exception as e:
        logger.error(f"Error looking up session {session_id}: {e}")
    
    return None

@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {"status": "healthy", "service": "vnc-bridge"}

@app.get("/vnc/{session_id}")
async def vnc_viewer(session_id: str):
    """Serve noVNC viewer page"""
    html_content = f"""
    <!DOCTYPE html>
    <html>
    <head>
        <title>VNC Viewer - Session {session_id}</title>
        <meta charset="utf-8">
        <style>
            body {{ margin: 0; padding: 0; overflow: hidden; }}
            #screen {{ position: fixed; top: 0; left: 0; width: 100%; height: 100%; }}
        </style>
    </head>
    <body>
        <div id="screen">
            <!-- Import noVNC -->
            <script type="module">
                import RFB from 'https://cdn.jsdelivr.net/npm/@novnc/novnc@1.4.0/core/rfb.js';
                
                const sessionId = '{session_id}';
                const wsUrl = `${{window.location.protocol === 'https:' ? 'wss' : 'ws'}}://${{window.location.host}}/vnc/${{sessionId}}/websocket`;
                
                // Create RFB client
                const rfb = new RFB(document.getElementById('screen'), wsUrl);
                
                // Configure RFB
                rfb.scaleViewport = true;
                rfb.resizeSession = false;
                
                // Event handlers
                rfb.addEventListener('connect', () => {{
                    console.log('Connected to VNC server');
                }});
                
                rfb.addEventListener('disconnect', (e) => {{
                    console.log('Disconnected from VNC server:', e.detail);
                    if (e.detail.clean) {{
                        document.getElementById('screen').innerHTML = '<div style="text-align: center; padding: 20px;">Session ended</div>';
                    }} else {{
                        document.getElementById('screen').innerHTML = '<div style="text-align: center; padding: 20px;">Connection lost. Trying to reconnect...</div>';
                        setTimeout(() => window.location.reload(), 3000);
                    }}
                }});
                
                rfb.addEventListener('credentialsrequired', () => {{
                    // No password for now
                    rfb.sendCredentials({{ password: '' }});
                }});
            </script>
        </div>
    </body>
    </html>
    """
    return HTMLResponse(content=html_content)

@app.websocket("/vnc/{session_id}/websocket")
async def vnc_websocket_proxy(websocket: WebSocket, session_id: str):
    """WebSocket proxy for VNC connections"""
    
    await websocket.accept()
    
    # Get desktop IP
    desktop_ip = await get_desktop_ip(session_id)
    if not desktop_ip:
        await websocket.send_text(json.dumps({"error": "Session not found"}))
        await websocket.close()
        return
    
    vnc_ws = None
    try:
        # Connect to desktop container's VNC WebSocket
        vnc_url = f"ws://{desktop_ip}:6080/websockify"
        logger.info(f"Connecting to VNC at {vnc_url} for session {session_id}")
        
        vnc_ws = await websockets.connect(vnc_url)
        
        # Proxy messages between client and VNC server
        async def client_to_vnc():
            try:
                while True:
                    data = await websocket.receive_bytes()
                    await vnc_ws.send(data)
            except WebSocketDisconnect:
                logger.info(f"Client disconnected for session {session_id}")
            except Exception as e:
                logger.error(f"Error in client->vnc proxy: {e}")
        
        async def vnc_to_client():
            try:
                async for message in vnc_ws:
                    if isinstance(message, bytes):
                        await websocket.send_bytes(message)
                    else:
                        await websocket.send_text(message)
            except Exception as e:
                logger.error(f"Error in vnc->client proxy: {e}")
        
        # Run both directions concurrently
        await asyncio.gather(client_to_vnc(), vnc_to_client())
        
    except Exception as e:
        logger.error(f"VNC proxy error for session {session_id}: {e}")
        try:
            await websocket.send_text(json.dumps({"error": str(e)}))
        except:
            pass
    finally:
        if vnc_ws:
            await vnc_ws.close()
        try:
            await websocket.close()
        except:
            pass

@app.get("/api/screenshot/{session_id}")
async def get_screenshot(session_id: str):
    """Get a screenshot from the VNC session"""
    
    desktop_ip = await get_desktop_ip(session_id)
    if not desktop_ip:
        raise HTTPException(status_code=404, detail="Session not found")
    
    try:
        # For now, return a placeholder
        # In production, this would capture from VNC
        return {"message": "Screenshot endpoint - to be implemented"}
    except Exception as e:
        logger.error(f"Screenshot error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/click/{session_id}")
async def send_click(session_id: str, x: int, y: int, button: str = "left"):
    """Send mouse click to VNC session"""
    
    desktop_ip = await get_desktop_ip(session_id)
    if not desktop_ip:
        raise HTTPException(status_code=404, detail="Session not found")
    
    try:
        # For now, return success
        # In production, this would send to VNC
        return {"success": True, "action": "click", "x": x, "y": y}
    except Exception as e:
        logger.error(f"Click error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/type/{session_id}")
async def send_type(session_id: str, text: str):
    """Send keyboard input to VNC session"""
    
    desktop_ip = await get_desktop_ip(session_id)
    if not desktop_ip:
        raise HTTPException(status_code=404, detail="Session not found")
    
    try:
        # For now, return success
        # In production, this would send to VNC
        return {"success": True, "action": "type", "text": text}
    except Exception as e:
        logger.error(f"Type error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8080)