backend/requirements-minimal.txt
# Minimal requirements for VNC bridge
fastapi==0.109.0
uvicorn[standard]==0.25.0
boto3==1.34.14
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
Pillow==10.2.0
vncdotool==1.2.0